refactor(core): split parser/grammar.rs into submodules (#516)
Break the 1,775-line grammar.rs into a grammar/ directory module:
mod.rs keeps the shared parser type aliases, ParserContext, and the
build_parser entry point; statements.rs, patterns.rs, and
expressions.rs host the statement, oracle-pattern, and
expression/literal combinators respectively, mirroring the eval-side
layout in abyss-interpreter. Items are pub(super) and shared between
siblings via private glob re-exports; build_parser remains the only
externally visible entry point. Pure code move.

Closes #508
